struggling with api loading
i literally forgot how to do this, I'm able to access the api
however when i return a function it gives me a promised object.
idk if i should return a state with the pokemon list or power through
to get the url through the promised object. Both are viable
the extra url object list feels easy but could lead to more bugs later
on