didn't work, it's all one pokemon(the last one)
I think what i'm going to do is instead of getting a new url on redner.
I would just update the url values in a pokemon object first on render.
And i think this is also going to save performance because I think
the component is rerendering all the urls every single time a component
gets updated.

And it make sense structurally because I only need to get the urls once