cheated using claude but i figured it out
1. jsx doesn't allow for in syntax. because jsx evaulates expressions
not statemnets so i must use the map.keys

2.there was a syntax error. I tried allPokemond.[dyn var].url
which is actually the correct way. HOWEVER while [] is for dynamic or
stored in a variable, I weas mixing the two syntaxes with the "." since
that "." is only for accessing a specific property name. and therefore
was causing an error.
I used claude to figure out this syntax error

next steps:
1. click + randomize function. This sets up an event listener for each
card.
2. create randomize function after setting up event listener