able to show clicks
now next step is to randomize when clicking.
and then after that I have to pass a prop. which shows the function
to check whether it's in the state of already clicked(i'll do a console)
if it's already clicked, and then reset the state, this funct is a prop