fixed clicked bug and enabled t/f statement
the bug issue was in the setClicked I was doing () instead of ([])
which is important because in setClicked i'm setting a new value
while also carrying the old values.
While this is differerent for objects, in array you're just copying
everything.

Unless, and I'm sure I can make the code more efficient, is you push
values into a new state and copy it over. I'm sure i can do that and
it makes the code more readable, but rn we'll just deal with this since
it's good enough